+++ app/src/main/java/com/withapp/coffeememo/di/search/bean/use_case/DeleteFilterBeanInputDataUseCaseModule.kt
@@ -1,6 +1,6 @@
 package com.withapp.coffeememo.di.search.bean.use_case
 
-import com.withapp.coffeememo.search.bean.domain.iterator.DeleteFilterBeanInputDataIterator
+import com.withapp.coffeememo.search.bean.domain.interactor.DeleteFilterBeanInputDataIterator
 import com.withapp.coffeememo.search.bean.domain.use_case.DeleteFilterBeanInputDataUseCase
 import dagger.Binds
 import dagger.Module
