+++ app/src/main/java/com/withapp/coffeememo/di/search/bean/use_case/FreeWordSearchBeanUseCaseModule.kt
@@ -1,6 +1,6 @@
 package com.withapp.coffeememo.di.search.bean.use_case
 
-import com.withapp.coffeememo.search.bean.domain.iterator.FreeWordSearchIterator
+import com.withapp.coffeememo.search.bean.domain.interactor.FreeWordSearchIterator
 import com.withapp.coffeememo.search.bean.domain.use_case.FreeWordSearchUseCase
 import dagger.Binds
 import dagger.Module
