+++ app/src/main/java/com/withapp/coffeememo/di/search/bean/use_case/GetAllBeanUseCaseModule.kt
@@ -1,6 +1,6 @@
 package com.withapp.coffeememo.di.search.bean.use_case
 
-import com.withapp.coffeememo.search.bean.domain.iterator.GetAllBeanIterator
+import com.withapp.coffeememo.search.bean.domain.interactor.GetAllBeanIterator
 import com.withapp.coffeememo.search.bean.domain.use_case.GetAllBeanUseCase
 import dagger.Binds
 import dagger.Module
