+++ app/src/main/java/com/withapp/coffeememo/di/search/bean/use_case/GetFilterBeanInputDataUseCaseModule.kt
@@ -1,6 +1,6 @@
 package com.withapp.coffeememo.di.search.bean.use_case
 
-import com.withapp.coffeememo.search.bean.domain.iterator.GetFilterBeanInputDataIterator
+import com.withapp.coffeememo.search.bean.domain.interactor.GetFilterBeanInputDataIterator
 import com.withapp.coffeememo.search.bean.domain.use_case.GetFilterBeanOutputDataUseCase
 import dagger.Binds
 import dagger.Module
