+++ app/src/main/java/com/withapp/coffeememo/di/search/bean/use_case/SetFilterBeanInputDataUseCaseModule.kt
@@ -1,6 +1,6 @@
 package com.withapp.coffeememo.di.search.bean.use_case
 
-import com.withapp.coffeememo.search.bean.domain.iterator.SetFilterBeanInputDataIterator
+import com.withapp.coffeememo.search.bean.domain.interactor.SetFilterBeanInputDataIterator
 import com.withapp.coffeememo.search.bean.domain.use_case.SetFilterBeanInputDataUseCase
 import dagger.Binds
 import dagger.Module
