+++ app/src/main/java/com/withapp/coffeememo/di/search/bean/use_case/SortBeanUseCaseModule.kt
@@ -1,6 +1,6 @@
 package com.withapp.coffeememo.di.search.bean.use_case
 
-import com.withapp.coffeememo.search.bean.domain.iterator.SortBeanInteractor
+import com.withapp.coffeememo.search.bean.domain.interactor.SortBeanInteractor
 import com.withapp.coffeememo.search.bean.domain.use_case.SortBeanUseCase
 import dagger.Binds
 import dagger.Module
