+++ app/src/main/java/com/withapp/coffeememo/di/search/bean/use_case/UpdateBeanFavoriteUseCaseModule.kt
@@ -1,6 +1,6 @@
 package com.withapp.coffeememo.di.search.bean.use_case
 
-import com.withapp.coffeememo.search.bean.domain.iterator.UpdateFavoriteBeanInterator
+import com.withapp.coffeememo.search.bean.domain.interactor.UpdateFavoriteBeanInterator
 import com.withapp.coffeememo.search.bean.domain.use_case.UpdateFavoriteBeanUseCase
 import dagger.Binds
 import dagger.Module
